requirements.txt
nepali_datetime==1.0.8.4
pandas==3.0.2
Pillow==12.2.0
plotly==6.7.0
streamlit==1.57.0
